[client] rename module
