+++ pycti/entities/stix_cyber_observable/opencti_stix_cyber_observable_properties.py
